requirements.txt
asgiref==3.8.1
attrs==23.2.0
boto3==1.34.150
botocore==1.34.150
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
deepl==1.18.0
diff-match-patch==20230430
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.2
django-import-export==4.1.1
django-modeltranslation==0.19.5
django-restful-translator==0.10.0
django-rosetta==0.10.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-core==2.4.1
google-cloud-translate==3.15.4
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.65.1
idna==3.7
inflection==0.5.1
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
nltk==3.8.1
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
polib==1.2.0
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rpds-py==0.19.0
rsa==4.9
s3transfer==0.10.2
scipy==1.14.0
setuptools==71.0.3
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
tqdm==4.66.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
